// src/app/features/inventario/models/requests/crear-producto-request.model.ts
export interface CrearProductoRequest {
  nombre: string;
  descripcion: string;
  precio: number;
  peso: number;
  requierePreparacion: boolean;
  categoriaId: string | null;
  nuevaCategoriaNombre: string;
  buffetId: string;
  stockActual: number;
  clasificacionesSaludIds: string[];
  tiposIds: null;
  urlImagen?: string | null;
}